Accommodation 필드 에러 수정 - reviewCount가 정상적으로 실행되지 않고 null 값이 나오던 문제를 수정 - Options 필드 추가

+++ BE/airbnb/src/main/resources/schema.sql
@@ -12,8 +12,8 @@ CREATE TABLE IF NOT EXISTS `accommodation`
     `options`    varchar(640),
     image        varchar(300),
     isLike       TINYINT(1),
-    `latitude`     DOUBLE(10,10),
-    `longitude`    DOUBLE(10,10)
+    `latitude`     DOUBLE(20,15),
+    `longitude`    DOUBLE(20,15)
 ) DEFAULT CHARSET = utf8;
 
 DESC accommodation;
